simplify shufti and provide arch-specific block functions
